pyproject.toml
[project]
name = "synthetic-text-watermarking"
version = "0.2.0"
description = "Demo of current approaches to watermarking synthetic text."
readme = "README.md"
authors = [
    {name = "Tom Potter", email = "tom.potter@bbc.co.uk"},
]
requires-python = ">=3.13"

dependencies = [
    "accelerate>=1.12.0",
    "c2pa-python>=0.28.0",
    "c2pa-text>=1.0.3",
    "cryptography>=46.0.5",
    "datasets>=4.5.0",
    "encypher-ai>=3.0.6",
    "flask>=3.1.2",
    "matplotlib>=3.10.8",
    "requests>=2.32.5",
    "ruff>=0.15.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.54.0",
    "tensorflow>=2.20.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.2.0",
    "watchdog>=6.0.0",
]

[project.scripts]
c2pa_demo = "synthetic_text_watermarking.c2pa.demo:main"
synthid_demo = "synthetic_text_watermarking.synthid.watermark:main"

[build-system]
requires = ["setuptools>=78.1.0", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"
